Fixed relationship between Allocation and Bill
