+++ backend/src/schemas/bill.graphql
@@ -32,5 +32,5 @@ type Mutation {
 }
 
 type Query {
-  bills(houseid: String!): [Bill]
+  bill(billid: String!): Bill
 }
