+++ backend/src/schemas/wishGroup.graphql
@@ -9,20 +9,20 @@ type WishGroup {
   color: String
   wishes: [Wish]
 }
-input EditWishGroupTitle{
-    houseID: String!
-    title: String!
-    updatedTitle: String!
+input EditWishGroupTitle {
+  houseID: String!
+  title: String!
+  updatedTitle: String!
 }
-input EditWishGroupColor{
-    houseID: String!
-    color: String!
-    updatedColor: String!
+input EditWishGroupColor {
+  houseID: String!
+  color: String!
+  updatedColor: String!
 }
 type Mutation {
   createWishGroup(wishgroup: WishGroupInput!): WishGroup
   deleteWishGroup(title: String!, houseid: String!): WishGroup
-  editGroupTitle(editTitleInput: EditWishGroup): WishGroup
+  editGroupTitle(editTitleInput: EditWishGroupTitle): WishGroup
   editGroupColor(editColorInput: EditWishGroupColor): WishGroup
 }
 type Query {
